quell warnings (part 2)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21127 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/FontManager.cxx
@@ -810,7 +810,7 @@ void FontManager::getPulseColor(const fvec4& color, fvec4& pulseColor) const
   // rate is how fast it should pulsate (smaller is faster)
   float pulseRate = BZDBCache::pulseRate;
 
-  float pulseFactor = fmodf(pulseTime, pulseRate) - (pulseRate * 0.5f);
+  float pulseFactor = fmodf((float)pulseTime, pulseRate) - (pulseRate * 0.5f);
   pulseFactor = fabsf(pulseFactor) / (pulseRate/2.0f);
   pulseFactor = pulseDepth * pulseFactor + (1.0f - pulseDepth);
